# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aplicativo MiauChei",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Local"
        }
    ],
    "paths": {
...
        "/pets-sighted": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "summary": "Pets avistados",
                "description": "Retorna lista de pets avistados",
                "operationId": "bde18f6a03d97baae39c97e0686329aa",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pet-sightings/{id}": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "summary": "Avistamentos do pet",
                "description": "Retorna os avistamentos do pet conforme id passado",
                "operationId": "007fbf338e3e0179a431922a5817c117",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pet",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pets-sighted-store": {
            "post": {
                "tags": [
                    "Pets"
                ],
                "summary": "Cadastrar avistamentos",
                "description": "Retorna dados do avistamento",
                "operationId": "d03f7d502c306b1b75b6d39734d91763",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data_sighted": {
                                        "type": "string"
                                    },
                                    "pet_id": {
                                        "type": "string"
                                    },
                                    "last_seen": {
                                        "type": "string"
                                    },
                                    "user_pet": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "data_sighted": "05/04/2022",
                                    "pet_id": "345",
                                    "last_seen": "Rua teste",
                                    "user_pet": true
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile": {
            "get": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Perfil do Usuário",
                "description": "Retorno do perfil do Usuário",
                "operationId": "5d0deb22f4dd2c57642ce790c940c21b",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Atualização do Perfil",
                "description": "Retorna dados atualizados",
                "operationId": "60b23a2c311e31aca76330467a2b88e5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "byte"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Armando",
                                    "phone": "999999999",
                                    "password": "12345678",
                                    "photo": "U3dhZ2dlciByb2Nrcw=="
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}